+++ src/Kraken.WebSockets.Sample/Program.cs
@@ -1,17 +1,29 @@
-﻿using System;
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.IO;
 using System.Threading.Tasks;
+using Kraken.WebSockets.Authentication;
 using Kraken.WebSockets.Logging;
 using Kraken.WebSockets.Messages;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Serilog;
 
 namespace Kraken.WebSockets.Sample
 {
+    [ExcludeFromCodeCoverage]
     class Program
     {
+        private static AuthToken token;
 
-        static void Main(string[] args)
+        static async Task Main(string[] args)
         {
+            var configuration = new ConfigurationBuilder()
+                .SetBasePath(Directory.GetCurrentDirectory())
+                .AddJsonFile("appsettings.json", true, true)
+                .AddEnvironmentVariables()
+                .Build();
+
             var logger = new LoggerConfiguration()
                 .MinimumLevel.Verbose()
                 .WriteTo.Console()
@@ -21,7 +33,16 @@ namespace Kraken.WebSockets.Sample
                 .AddKrakenWebSockets()
                 .AddSerilog(logger);
 
-            using (var client = KrakenApi.ClientFactory.Create("wss://ws.kraken.com"))
+            var krakenApi = new KrakenApi()
+                .ConfigureAuthentication(
+                    configuration.GetValue<string>("API_URL"),
+                    configuration.GetValue<string>("API_KEY"),
+                     configuration.GetValue<string>("API_SECRET"))
+                .ConfigureWebsocket("wss://ws.kraken.com");
+
+            token = await krakenApi.AuthenticationClient.GetWebsocketToken();
+
+            using (var client = krakenApi.BuildClient())
             {
                 Task.Run(() => RunKraken(client));
                 do
@@ -43,6 +64,8 @@ namespace Kraken.WebSockets.Sample
             client.SpreadReceived += (sender, e) => Console.WriteLine($"Spread received");
             client.BookSnapshotReceived += (sender, e) => Console.WriteLine($"BookSnapshot received");
             client.BookUpdateReceived += (sender, e) => Console.WriteLine($"BookUpdate received");
+            client.OwnTradesReceived += (sender, e) => Console.WriteLine($"OwnTrades received");
+            client.OpenOrdersReceived += (sender, e) => Console.WriteLine($"OpenOrders received");
 
             await client.ConnectAsync();
 
@@ -56,6 +79,8 @@ namespace Kraken.WebSockets.Sample
             };
 
             await client.SubscribeAsync(new Subscribe(new[] { Pair.XBT_EUR }, new SubscribeOptions(SubscribeOptionNames.All)));
+            await client.SubscribeAsync(new Subscribe(null, new SubscribeOptions(SubscribeOptionNames.OwnTrades, token.Token)));
+            await client.SubscribeAsync(new Subscribe(null, new SubscribeOptions(SubscribeOptionNames.OpenOrders, token.Token)));
         }
     }
 }