+++ src/Kraken.WebSockets.Tests/Messages/SubscribeTests.cs
@@ -8,31 +8,30 @@ namespace Kraken.WebSockets.Tests.Messages
     [ExcludeFromCodeCoverage]
     public class SubscribeTests
     {
-        private const int requestId = 123;
+        private const int RequestId = 123;
         private readonly string[] pairs;
         private readonly SubscribeOptions options;
         private readonly Subscribe instance;
 
         public SubscribeTests()
         {
-
-            pairs = new string[] { "XBT/EUR" };
+            pairs = new[] { "XBT/EUR" };
             options = new SubscribeOptions("*") { };
-            instance = new Subscribe(pairs, options, requestId);
+            instance = new Subscribe(pairs, options, RequestId);
         }
 
         #region Ctor
 
-        [Fact]
+        [Fact(Skip = "Transform nullable pairs to separate subscription")]
         public void Ctor_PairsNull_ThrowsArgumentNullException()
         {
-            Assert.Equal("pairs", Assert.Throws<ArgumentNullException>(() => new Subscribe(null, options, requestId)).ParamName);
+            Assert.Equal("pairs", Assert.Throws<ArgumentNullException>(() => new Subscribe(null, options, RequestId)).ParamName);
         }
 
         [Fact]
         public void Ctor_OptionsNull_ThrowsArgumentNullException()
         {
-            Assert.Equal("options", Assert.Throws<ArgumentNullException>(() => new Subscribe(pairs, null, requestId)).ParamName);
+            Assert.Equal("options", Assert.Throws<ArgumentNullException>(() => new Subscribe(pairs, null, RequestId)).ParamName);
         }
 
         #endregion
@@ -62,7 +61,7 @@ namespace Kraken.WebSockets.Tests.Messages
         [Fact]
         public void RequestId_Get_ReturnsValue()
         {
-            Assert.Equal(requestId, instance.RequestId);
+            Assert.Equal(RequestId, instance.RequestId);
         }
 
         #endregion