+++ src/Kraken.WebSockets.Tests/TestSocketMessages.cs
@@ -163,5 +163,17 @@ namespace Kraken.WebSockets.Tests
         }
 
         #endregion
+
+        #region OwnTradesReceived
+
+        public static string OwnTradesMessage => File.ReadAllText("data/ownTrades.json");
+
+        #endregion
+
+        #region OpenOrders
+
+        public static string OpenOrdersMessage => File.ReadAllText("data/openOrders.json");
+
+        #endregion
     }
 }