+++ src/Kraken.WebSockets/Authentication/AuthToken.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Kraken.WebSockets.Authentication
+{
+    public class AuthToken
+    {
+        public string Token{ get; set; }
+
+        public decimal? Expires { get; set; }
+    }
+}