+++ src/Kraken.WebSockets/Authentication/IAuthenticationClient.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Kraken.WebSockets.Authentication
+{
+    public interface IAuthenticationClient
+    {
+        Task<AuthToken> GetWebsocketToken();
+    }
+}