+++ src/Kraken.WebSockets/IKrakenApiClient.cs
@@ -69,6 +69,16 @@ namespace Kraken.WebSockets
         event EventHandler<KrakenDataEventArgs<BookUpdateMessage>> BookUpdateReceived;
 
         /// <summary>
+        /// Occurs when own trades information was received.
+        /// </summary>
+        event EventHandler<KrakenPrivateEventArgs<OwnTradesMessage>> OwnTradesReceived;
+
+        /// <summary>
+        /// Occurs when open orders information was received.
+        /// </summary>
+        event EventHandler<KrakenPrivateEventArgs<OpenOrdersMessage>> OpenOrdersReceived;
+
+        /// <summary>
         /// Connects to the websocket endpoint.
         /// </summary>
         /// <returns></returns>