+++ src/Kraken.WebSockets/Messages/SubscribeOptionNames.cs
@@ -5,7 +5,10 @@ namespace Kraken.WebSockets.Messages
     /// <summary>
     /// Subscribe option names.
     /// </summary>
-    /// <remarks>ticker|ohlc|trade|book|spread|*</remarks>
+    /// <remarks>
+    /// public: ticker|ohlc|trade|book|spread|*
+    /// private: ownTrades (you need to provide an authentication token)
+    /// </remarks>
     public static class SubscribeOptionNames
     {
         /// <summary>
@@ -44,6 +47,17 @@ namespace Kraken.WebSockets.Messages
         /// </summary>
         public const string Spread = "spread";
 
+        /// <summary>
+        /// Own trades, on subscription last 50 trades for the user will be sent, followed by new trades.
+        /// </summary>
+        public const string OwnTrades = "ownTrades";
+
+        /// <summary>
+        /// Open orders, initial snapshot will provide list of all open orders and then any updates to
+        /// the open orders list will be sent.
+        /// </summary>
+        public const string OpenOrders = "openOrders";
+
         internal static IEnumerable<string> AllowedNames
         {
             get
@@ -55,6 +69,17 @@ namespace Kraken.WebSockets.Messages
                 yield return Trade;
                 yield return Book;
                 yield return Spread;
+                yield return OwnTrades;
+                yield return OpenOrders;
+            }
+        }
+
+        internal static IEnumerable<string> PrivateNames
+        {
+            get
+            {
+                yield return OwnTrades;
+                yield return OpenOrders;
             }
         }
     }